Chore: bonus suffix 추가 및 makefile 작업
