+++ Makefile
@@ -1,38 +1,44 @@
 NAME =	miniRT
 
-SRCS =	src/mlx/mlx_utils.c \
-		src/print/print.c \
-		src/scene/canvas.c \
-		src/scene/object_create.c \
-		src/scene/render_scene.c \
-		src/scene/scene.c \
-		src/trace/hit/hit_sphere.c \
-		src/trace/hit/hit_plane.c \
-		src/trace/hit/hit_cylinder.c \
-		src/trace/hit/hit_cy_base.c \
-		src/trace/hit/hit.c \
-		src/trace/hit/normal.c \
-		src/trace/ray/phong_lighting.c \
-		src/trace/ray/ray.c \
-		src/trace/ray/reflect.c \
-		src/utils/color/color3_utils.c \
-		src/utils/object/object_utils.c \
-		src/utils/vec3/vec3_utils.c \
-		src/utils/vec3/vec3_constructor.c \
-		src/utils/vec3/vec3_divide.c \
-		src/utils/vec3/vec3_minus.c \
-		src/utils/vec3/vec3_mult.c \
-		src/utils/vec3/vec3_plus.c \
-		src/utils/vec3/vec3_length.c \
-		src/parsing/process_file.c \
-		src/parsing/split_line.c \
-		src/parsing/double_conversion.c \
-		src/parsing/parsing_utils.c \
-		src/parsing/get_xyz.c \
-		src/parsing/put_acl.c \
-		src/parsing/put_object.c \
-		src/main.c
-OBJS = ${SRCS:.c=.o}
+RAW_SRCS =	src/mlx/mlx_utils \
+			src/print/print \
+			src/scene/canvas \
+			src/scene/object_create \
+			src/scene/render_scene \
+			src/scene/scene \
+			src/trace/hit/hit_sphere \
+			src/trace/hit/hit_plane \
+			src/trace/hit/hit_cylinder \
+			src/trace/hit/hit_cy_base \
+			src/trace/hit/hit \
+			src/trace/hit/normal \
+			src/trace/ray/phong_lighting \
+			src/trace/ray/ray \
+			src/trace/ray/reflect \
+			src/utils/color/color3_utils \
+			src/utils/object/object_utils \
+			src/utils/vec3/vec3_utils \
+			src/utils/vec3/vec3_constructor \
+			src/utils/vec3/vec3_divide \
+			src/utils/vec3/vec3_minus \
+			src/utils/vec3/vec3_mult \
+			src/utils/vec3/vec3_plus \
+			src/utils/vec3/vec3_length \
+			src/parsing/process_file \
+			src/parsing/split_line \
+			src/parsing/double_conversion \
+			src/parsing/parsing_utils \
+			src/parsing/get_xyz \
+			src/parsing/put_acl \
+			src/parsing/put_object \
+			src/main
+MAN_SRCS = $(addsuffix .c, ${RAW_SRCS})
+MAN_OBJS = ${MAN_SRCS:.c=.o}
+
+BONUS_DIR = bonus
+BONUS_SRCS = $(addsuffix $(addprefix ${BONUS_DIR}/, ${RAW_SRCS}), _bonus.c)
+BONUS_SRCS += ${BONUS_DIR}/src/trace/hit/hit_cone_bonus.c
+BONUS_OBJS = ${BONUS_SRCS:.c=.o}
 
 MLX_DIR = minilibx
 MLX_LIB = libmlx.dylib
@@ -42,22 +48,36 @@ LIBFT_DIR = libft
 CC = cc
 CFLAGS = -Wall -Wextra -Werror -g
 LFLAGS = -L$(LIBFT_DIR) -lft -L$(MLX_DIR) -lmlx -framework OpenGL -framework AppKit
-INC = -I include -I ${LIBFT_DIR} -I ${MLX_DIR}
+INC = -I ${LIBFT_DIR} -I ${MLX_DIR}
 RM = rm -f
 
 all:	${NAME}
 
-${NAME}: ${OBJS}
+bonus:
+	@make bonus=1 all
+
+ifdef bonus
+	SRCS = ${BONUS_SRCS}
+	OBJS = ${BONUS_OBJS}
+	INC += -I ${BONUS_DIR}/include
+else
+	SRCS = ${MAN_SRCS}
+	OBJS = ${MAN_OBJS}
+	INC += -I include
+endif
+
+$(NAME): ${OBJS}
 		@make bonus -C libft
 		@make -C minilibx
 		@cp $(MLX_DIR)/$(MLX_LIB) .
 		@${CC} ${CFLAGS} ${LFLAGS} -o ${NAME} ${OBJS}
 
 %.o: %.c
+		echo $<
 		@${CC} ${CFLAGS} ${INC} -c $< -o ${<:.c=.o}
 
 clean:
-		${RM} ${OBJS}
+		${RM} ${OBJS} ${BONUS_OBJS}
 		@make clean -C libft
 		@make clean -C minilibx
 
@@ -68,3 +88,5 @@ fclean:	clean
 		${RM} ${MLX_LIB}
 
 re:		fclean all
+
+.PHONY:	all clean fclean re bonus
