+++ bonus/src/parsing/put_object_bonus.c
