+++ bonus/src/scene/render_scene_bonus.c
