+++ bonus/src/trace/hit/hit_bonus.c
