+++ bonus/src/trace/hit/hit_cylinder_bonus.c
