+++ bonus/src/trace/hit/normal_bonus.c
