+++ bonus/src/trace/ray/phong_lighting_bonus.c
