+++ bonus/src/trace/ray/ray_bonus.c
