+++ bonus/src/utils/object/object_utils_bonus.c
