+++ bonus/src/utils/vec3/vec3_mult_bonus.c
